+++ src/sass/elements/_header.scss
@@ -15,6 +15,8 @@
     border: solid 2px lightgrey;
     background-color: white;
     right: 0;
+    width: 30px;
+    height: 30px;
   }
 }
